added a feature which verify ip

+++ app/controllers/weather_controller.rb
@@ -3,6 +3,12 @@ class WeatherController < ApplicationController
 	skip_before_action :verify_authenticity_token
 
 	def update
+		
+		unless request.remote_ip == '127.0.0.1'
+			respond_to {|format| format.json { render json: {msg: 'illegal ip', status: 1 } } }
+			return
+		end
+
 		obj = JSON.parse(params[:data])
 	
 		Weather.delete_all
@@ -18,6 +24,7 @@ class WeatherController < ApplicationController
 		end
 
 	    respond_to {|format| format.json { render json: {msg: "has been inserted", status: 0 } } }
+	    
 
 	end
